+++ WeatherForecast/WeatherForecast/Controller/MainWeatherTableViewController.swift
@@ -1,0 +1,54 @@
+//
+//  WeatherForecast - ViewController.swift
+//  Created by yagom.
+//  Copyright © yagom. All rights reserved.
+//
+
+import UIKit
+
+class MainWeatherTableViewController: UITableViewController {
+    private let weatherDataViewModel: WeatherDataViewModel
+        
+    init(weatherDataViewModel: WeatherDataViewModel) {
+        self.weatherDataViewModel = weatherDataViewModel
+        super.init(nibName: nil, bundle: nil)
+    }
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        let backgroundImageView = UIImageView(image: UIImage(named: "NightStreet"))
+        backgroundImageView.contentMode = .scaleAspectFill
+        
+        weatherDataViewModel.setUpWeatherData {
+            self.tableView.reloadData()
+        }
+    }
+}
+
+// MARK: - TableView DataSource
+extension MainWeatherTableViewController {
+    override func numberOfSections(in tableView: UITableView) -> Int {
+        return 1
+    }
+    
+    override func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
+        return 0.0
+    }
+    
+    override func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
+        return nil
+    }
+    
+    
+    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return weatherDataViewModel.intervalWeatherInfos.count
+    }
+    
+    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        return UITableViewCell()
+    }
+}
